+++ darts/tests/test_local_forecasting_models.py
@@ -7,7 +7,10 @@ from .base_test_class import DartsBaseTestClass
 from ..timeseries import TimeSeries
 from ..utils import timeseries_generation as tg
 from ..metrics import mape
-from ..models import NaiveSeasonal, ExponentialSmoothing, ARIMA, Theta, FourTheta, FFT, VARIMA
+from ..models import (
+    NaiveSeasonal, ExponentialSmoothing, ARIMA, Theta, FourTheta, FFT, VARIMA,
+    RandomForest, LinearRegressionModel
+)
 from ..utils.utils import SeasonalityMode, TrendMode, ModelMode
 from ..logging import get_logger
 
@@ -28,6 +31,8 @@ models = [
     (FourTheta(season_mode=SeasonalityMode.ADDITIVE), 14.2),
     (FFT(trend='poly'), 11.4),
     (NaiveSeasonal(), 32.4),
+    (LinearRegressionModel(lags=12), 11.0),
+    (RandomForest(lags=12, n_estimators=200, max_depth=3), 15.0),
 ]
 # forecasting models with exogenous variables support
 multivariate_models = [
@@ -35,7 +40,7 @@ multivariate_models = [
     (VARIMA(1, 1, 1), 57.0),
 ]
 
-extended_models = [ARIMA()]
+extended_models = [ARIMA(), LinearRegressionModel(lags=4, lags_exog=2), RandomForest(lags=4, lags_exog=2)]
 
 
 try:
@@ -92,7 +97,7 @@ class LocalForecastingModelsTestCase(DartsBaseTestClass):
             model.fit(self.ts_pass_train)
             prediction = model.predict(len(self.ts_pass_val))
             current_mape = mape(prediction, self.ts_pass_val)
-            self.assertTrue(current_mape < max_mape, "{} model exceeded the maximum MAPE of {}."
+            self.assertTrue(current_mape < max_mape, "{} model exceeded the maximum MAPE of {}. "
                             "with a MAPE of {}".format(str(model), max_mape, current_mape))
 
     def test_multivariate_models_performance(self):
@@ -101,7 +106,7 @@ class LocalForecastingModelsTestCase(DartsBaseTestClass):
             model.fit(self.ts_ice_heater_train)
             prediction = model.predict(len(self.ts_ice_heater_val))
             current_mape = mape(prediction, self.ts_ice_heater_val)
-            self.assertTrue(current_mape < max_mape, "{} model exceeded the maximum MAPE of {}."
+            self.assertTrue(current_mape < max_mape, "{} model exceeded the maximum MAPE of {}. "
                             "with a MAPE of {}".format(str(model), max_mape, current_mape))
 
     def test_multivariate_input(self):